+++ packages/database/src/models/fhir/FhirServiceRequest.ts
@@ -5,9 +5,8 @@ import {
   FHIR_INTERACTIONS,
   SERVICE_REQUEST_CATEGORY_CODES,
   SERVICE_REQUEST_PERMISSION_NOUNS,
-  FHIR_INTEGRATION_VERB,
-  FHIR_INTEGRATION_PERMISSIONS,
 } from '@tamanu/constants';
+import { hasFhirPermission } from '@tamanu/shared/permissions/hasFhirPermission';
 import { FhirResource } from './Resource';
 import type { Models } from '../../types/model';
 import {
@@ -19,19 +18,6 @@ import {
   searchParameters,
   shouldForceRematerialise,
 } from '../../utils/fhir/ServiceRequest';
-
-function hasFhirPermission(ability: Ability, verb: string, noun: string): boolean {
-  if (ability.can(verb, noun)) return true;
-
-  for (const [type, config] of Object.entries(FHIR_INTEGRATION_PERMISSIONS)) {
-    const hasFullAccess = ability.can(FHIR_INTEGRATION_VERB, type);
-    const hasVerbAccess = ability.can(verb, type);
-    if (!hasFullAccess && !hasVerbAccess) continue;
-    if (verb === 'read' && config.read.includes(noun)) return true;
-    if (verb === 'write' && config.write.includes(noun)) return true;
-  }
-  return false;
-}
 
 function getAllowedCategories(ability: Ability): string[] {
   const categories: string[] = [];
@@ -129,7 +115,9 @@ export class FhirServiceRequest extends FhirResource {
     ability: Ability,
   ): Record<string, any> {
     const allowedCategories = getAllowedCategories(ability);
-    if (allowedCategories.length === 0) return query;
+    if (allowedCategories.length === 0) {
+      return { ...query, where: { ...query.where, id: null } };
+    }
 
     const categoryConditions = allowedCategories.map(code =>
       Sequelize.literal(`category @> '[{"coding": [{"code": "${code}"}]}]'::jsonb`),
@@ -150,23 +138,6 @@ export class FhirServiceRequest extends FhirResource {
         ],
       },
     };
-  }
-
-  static checkRecordAccess(ability: Ability, record: FhirResource): void {
-    const allowedCategories = getAllowedCategories(ability);
-    if (allowedCategories.length === 0) {
-      throw new Error('No permission to read ServiceRequest');
-    }
-
-    const serviceRequest = record as FhirServiceRequest;
-    const recordCategories = serviceRequest.category ?? [];
-    const recordCodes = (recordCategories as any[]).flatMap((cat: any) =>
-      (cat.coding ?? []).map((coding: any) => coding.code),
-    );
-
-    if (!recordCodes.some(code => allowedCategories.includes(code))) {
-      throw new Error(`no ServiceRequest with id ${record.id}`);
-    }
   }
 
   async updateMaterialisation() {
